+++ projects/core/src/column.ts
@@ -9,11 +9,10 @@ export function makeTitle(name: string) {
   // insert a space before all caps
   return (
     name
-      .replace(/([A-Z])/g, ' $1')
-      // uppercase the first character
+      .replace(/([a-z])([A-Z])/g, '$1 $2')
+      // Uppercase the first character
       .replace(/^./, (str) => str.toUpperCase())
       .replace('Email', 'eMail')
-      .replace(' I D', ' ID')
   )
 }