+++ projects/core/src/remult-proxy.ts
@@ -215,16 +215,21 @@ export class RemultProxy implements Remult {
       get metadata() {
         return remultStatic.remultFactory().repo(...args).metadata
       },
-      query: (...args2) =>
+      query: (options: any) =>
         self
           .remultFactory()
           .repo(...args)
-          .query(...args2),
+          .query(options) as any,
       save: (args2: any) =>
         self
           .remultFactory()
           .repo(...args)
           .save(args2),
+      upsert: (args2: any) =>
+        self
+          .remultFactory()
+          .repo(...args)
+          .upsert(args2),
       update: (a: any, b: any) =>
         self
           .remultFactory()