+++ CHANGELOG.md
@@ -2,13 +2,59 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.27.21] TBD
+
+- **Added `upsert` method:**  
+  The `upsert` method allows inserting or updating an entity in a single operation. If an entity matching the `where` condition is found, it is updated; otherwise, a new entity is created. This can be used for a single entity or for batch operations with an array of options.
+
+  Example:
+
+  ```ts
+  // Single entity upsert
+  await taskRepo.upsert({
+    where: { title: 'task a' },
+    set: { completed: true },
+  })
+
+  // Batch upsert
+  await taskRepo.upsert([
+    { where: { title: 'task a' }, set: { completed: true } },
+    { where: { title: 'task b' }, set: { completed: true } },
+  ])
+  ```
+
+- Now you can get data and aggregate info with a single request using the `query` method:
+
+  ```ts
+  const result = await repo
+    .query({
+      where: { completed: false },
+      pageSize: 50,
+      aggregates: {
+        sum: ['salary'],
+        average: ['age'],
+      },
+    })
+    .paginator()
+  // Accessing the items from the first page
+  console.table(result.items)
+  // Accessing the aggregation results
+  console.log(result.aggregates.salary.sum) // Total salary sum
+  ```
+
+- Fixed issue with `updateMany` where the `set` had `ValueListType` or relation.
+- Fixed issue when updating relation id and relation in the same update, the last one will win
+- Fixed makeTitle to handle all caps text
+
 ## [0.27.20] 2024-10-16
 
-- Added Origin IndexedDb Storage to store entities in the front end
+- **Added Origin IndexedDb Storage** to store entities in the frontend:
+
   ```ts
   const db = new JsonDataProvider(new JsonEntityIndexedDbStorage())
   console.table(await repo(Task, db).find())
   ```
+
 - Fixed issue with required validation and relations
 
 ## [0.27.19] 2024-09-26